fingerprint: set dimlayer_hbm on finger up/down instead of auth/cancel When dimlayer_hbm is enabled, HWC ends up disabling all color transformations, including Night Light. Since we enable/disable dimlayer_hbm on authenticate/cancel, this means that whenever the UDFPS circle is present, Night Light ends up temporarily turning off. This is especially not helpful on the lock screen, where the UDFPS circle is active all the time.
Instead of turning on dimlayer_hbm during the entire time that the UDFPS
circle is showing, let's turn on dimlayer_hbm only when the user is
pressing down on the circle. In doing so, the color transformations will
be disabled for only when the user is touching the circle.

We will still disable dimlayer_hbm on cancel anyway, just in case that
wasn't already set to off by the time cancel() is called.

Change-Id: I9ed1a6b5839e39b8f70ff4af4ebe8d1bc66ffb98